# openapi/openapi.json
{
  "openapi" : "3.0.3",
  "info" : {
    "title" : "TodoApiJava API",
    "version" : "1.0.1"
  },
  "servers" : [ {
    "url" : "http://localhost:8080",
    "description" : "Auto generated value"
  }, {
    "url" : "http://0.0.0.0:8080",
    "description" : "Auto generated value"
  } ],
  "paths" : {
    "/todo/items" : {
      "get" : {
        "tags" : [ "Todo Resource" ],
        "responses" : {
          "200" : {
            "description" : "Returns a List of TodoItems",
            "content" : {
              "application/json" : {
                "schema" : {
                  "type" : "array",
                  "items" : {
                    "$ref" : "#/components/schemas/TodoItemJava"
                  }
                }
              }
            }
          }
        }
      },
      "post" : {
        "tags" : [ "Todo Resource" ],
        "requestBody" : {
          "content" : {
            "application/json" : {
              "schema" : {
                "$ref" : "#/components/schemas/TodoItemJava"
              }
            }
          }
        },
        "responses" : {
          "201" : {
            "description" : "TodoItem Created",
            "content" : {
              "application/json" : {
                "schema" : {
                  "$ref" : "#/components/schemas/TodoItemJava"
                }
              }
            }
          }
        }
      }
    },
    "/todo/items/{id}" : {
      "get" : {
        "tags" : [ "Todo Resource" ],
        "parameters" : [ {
          "name" : "id",
          "in" : "path",
          "required" : true,
          "schema" : {
            "format" : "int64",
            "type" : "integer"
          }
        } ],
        "responses" : {
          "200" : {
            "description" : "Returns a TodoItem",
            "content" : {
              "application/json" : {
                "schema" : {
                  "$ref" : "#/components/schemas/TodoItemJava"
                }
              }
            }
          },
          "404" : {
            "description" : "If Not Found"
          }
        }
      },
      "put" : {
        "tags" : [ "Todo Resource" ],
        "parameters" : [ {
          "name" : "id",
          "in" : "path",
          "required" : true,
          "schema" : {
            "format" : "int64",
            "type" : "integer"
          }
        } ],
        "requestBody" : {
          "content" : {
            "application/json" : {
              "schema" : {
                "$ref" : "#/components/schemas/TodoItemJava"
              }
            }
          }
        },
        "responses" : {
          "204" : {
            "description" : "Returns No Content"
          },
          "400" : {
            "description" : "If Request is Bad"
          },
          "404" : {
            "description" : "If Not Found"
          }
        }
      },
      "delete" : {
        "tags" : [ "Todo Resource" ],
        "parameters" : [ {
          "name" : "id",
          "in" : "path",
          "required" : true,
          "schema" : {
            "format" : "int64",
            "type" : "integer"
          }
        } ],
        "responses" : {
          "204" : {
            "description" : "Returns No Content"
          },
          "404" : {
            "description" : "If Not Found"
          }
        }
      }
    }
  },
  "components" : {
    "schemas" : {
      "TodoItemJava" : {
        "type" : "object",
        "properties" : {
          "id" : {
            "format" : "int64",
            "type" : "integer"
          },
          "name" : {
            "type" : "string"
          },
          "done" : {
            "type" : "boolean"
          }
        }
      }
    }
  }
}